Split telekinesis into cross-platform core and JVM transport modules
The HTTP model, client typeclasses and redirect loop now cross-compile to
Scala.js; `telekinesis.jvm` carries the `java.net.http` `Http.Backend`
(summoned by `import httpBackends.virtualMachine`) and the `coaxial`
domain-socket client, whose givens moved out of the companion objects.
