+++ lib/ethereal/src/core/ethereal.Runners.scala
@@ -46,6 +46,7 @@ import vacuous.*
 
 import errorDiagnostics.emptyDiagnostics
 import gastronomy.*, providers.javaStdlibProvider
+import httpBackends.virtualMachine
 import internetAccess.online
 import monotonous.*, alphabets.hexLowerCase
