+++ lib/obligatory/src/json/obligatory.JsonRpc.scala
@@ -54,6 +54,8 @@ import urticose.*
 import vacuous.*
 import zephyrine.*
 
+import httpBackends.virtualMachine
+
 
 object JsonRpc:
   private val promises: scm.HashMap[Text | Int, Promise[Json]] = scm.HashMap()
