+++ lib/orthodoxy/src/core/orthodoxy.Issuer.scala
@@ -51,6 +51,7 @@ import vacuous.*
 import zephyrine.*
 
 import errorDiagnostics.stackTracesDiagnostics
+import httpBackends.virtualMachine
 import queryParameters.arbitraryQueryParameter
 
 object Issuer:
